poc/requirements.txt
aws-cdk-lib>=2.170.0
constructs>=10.0.0,<11.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Synthetic data generation
faker>=19.0.0
boto3>=1.34.0

# Lambda function dependencies (for local testing)
botocore>=1.34.0

